requirements.txt
gradio>=4.0.0
fastapi>=0.100.0
uvicorn>=0.20.0
huggingface_hub>=0.20.0
pydantic>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
requests>=2.31.0
openai>=1.0.0
